Service Threads & Config
Allow thread count to be configured in config tool & make calculated thread count visible.
Change the system thread calculation to be smarter.  Instead of being hard coded to 10 it will increase based on the number of source connections and available processors with a min of 10 and max of 100 threads.  The thread count might not always be ideal, but users can override it if needed.